backend/requirements.txt
flask
flask-pymongo
flask-cors
bcrypt
pyjwt
gunicorn